dockerize demo app
