--- App/Template/base.html
@@ -1,5 +1,0 @@
-{% extends "bootstrap/base.html" %}
-
-{% block title %}Satellite Imagery{% endblock %}
-
-{% block content %}{% endblock %}
