+++ demo/README.md
@@ -1,0 +1,13 @@
+# Tensorflow serving
+
+[Tensorflow Serving REST Tutorial](https://www.tensorflow.org/tfx/tutorials/serving/rest_simple)
+
+## Getting started
+``` bash
+docker-compose up
+```
+
+``` bash
+(satellite_imagery)fiete@ubu:/ssd2/development/satellite_imagery/App/scripts$ python predict_image.py 
+Data: {"signature_name": "serving_default", "instances": ... 29275513, 0.9463414549827576, 0.9793103337287903]]]}
+```
