+++ demo/flask_server/app/template/base.html
@@ -1,0 +1,5 @@
+{% extends "bootstrap/base.html" %}
+
+{% block title %}Satellite Imagery{% endblock %}
+
+{% block content %}{% endblock %}
