+++ demo/flask_server/app/template/index.html
@@ -1,0 +1,17 @@
+{% extends "base.html" %}
+
+{% block content %}
+ <div class="container">
+   <h1>Upload File</h1>
+   <hr>
+   <form class="form-inline" action="{{ url_for('index') }}" method="post" enctype="multipart/form-data">
+     <div class="form-group">
+       <input type="file" name="file" class="btn btn-default">
+     </div>
+     <div class="form-group">
+       <input type="submit" value="Upload" class="btn btn-default">
+     </div>
+   </form>
+ </div>
+
+{% endblock %}
