+++ demo/flask_server/app/template/result.html
@@ -1,0 +1,14 @@
+{% extends "base.html" %}
+
+{% block content %}
+  <div class="container">
+    <h1>Imageplot</h1>
+    <img src={{result.img_path}} alt="img-rounded" height="100" width="100">
+    <hr>
+
+    <hr>
+    <a href="{{ url_for('index') }}" class="btn btn-default">Back</a>
+  </div>
+
+{% endblock %}
+
