+++ demo/flask_server/requirements
@@ -1,0 +1,7 @@
+flask
+flask_bootstrap
+matplotlib
+numpy
+requests
+opencv-python-headless
+pandas
